Ignore all status checks when performing automerge
Because renovate does not consider branch protection rules
and looks at all checks, automerge will fail if ANY test fails.

This is the only way to ignore checks that are failed but
are not required by branch protection rules.

To prevent merging without required checks, we MUST ensure
that balena-renovate[bot] does NOT have admin:write permissions
or any other settings that would allow it to bypass protections.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>